requirements.txt
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
docx==0.2.4
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
pep257==0.7.0
Pillow==8.2.0
pycodestyle==2.7.0
pydocstyle==6.1.1
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2021.1
requests==2.25.1
six==1.16.0
snowballstemmer==2.1.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
